codebase/requirements.txt
Flask==3.0.1
python-dotenv==1.0.0
requests==2.31.0
flask-cors==4.0.0
python-logging==0.4.9.6
pytest==7.4.3
gunicorn==21.2.0 